+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Cover `<AcceptedValuesField>` component with unit tests (UIDATIMP-692)
 * Cover `<MappedHeader>` component with unit tests (UIDATIMP-693)
 * Reuse `<JobsListAccordion>` component from `stripes-data-transfer-components` rep (UIDATIMP-574)
+* Reuse `<JobsLogs>` component from `stripes-data-transfer-components` rep (UIDATIMP-572)
 
 ### Bugs fixed:
 * Log filter dropdowns by job profile and user are not in alphabetical order (UIDATIMP-601)